* Moving logic from the controller to the manager classes * Setup spring bean configurations * Removed overidden versions from pom.xml * Added manager classes to hand R and Prediction Generation

+++ src/main/java/org/sjsu/cmpe/dao/impl/mapper/ZillowPropertyDetailStageMapper.java
@@ -14,7 +14,7 @@ public class ZillowPropertyDetailStageMapper implements RowMapper<ZillowProperty
         zillowPropertyDetailStage.setHomeDescription(rs.getString("home_description"));
         zillowPropertyDetailStage.setImageCount(rs.getInt("image_count"));
         zillowPropertyDetailStage.setPageViewCount(rs.getLong("page_view_count"));
-        zillowPropertyDetailStage.setZpid(rs.getLong("zpid"));
+        zillowPropertyDetailStage.setZpid(rs.getInt("zpid"));
         return zillowPropertyDetailStage;
     }
 }
